Moved SwipeRefreshLayout into UpdaterFragment and undid the associated changes to MainActivity. This fixes scrolling for the RecyclerViews and localizes the functionality to the UpdaterFragment.
